More text in the readme and changed the texture and mesh.
Texture now add a new function to create a texture from a program and a texture manager. also changed the other one to take the same parameters.

Also changed the mesh to be able to create a correcte quad and started the work for the same thing with the cube (not done yet).